+++ sim/game/cards/repository.go
@@ -794,6 +794,8 @@ var DM10 = map[string]match.CardConstructor{
 	"e46de1fa-e307-4119-ac71-69c95dc5e443": dm10.CoreCrashLizard,
 	"b9380b9c-894f-421b-b060-1450a4eb32cd": dm10.ElixiaPurebladeElemental,
 	"31683921-16fc-4d4c-bb77-3225d10f7366": dm10.UltimateDragon,
+	"8f0df729-fa1a-439e-a4a1-961650bcd192": dm10.NecrodragonBryzenaga,
+	"7a39340a-4601-4c34-8754-05554d49cbf4": dm10.BodaciousGiant,
 }
 
 // Promo is a map with all the card id's in the game and corresponding CardConstructor for promotional exclusive cards